requirements.txt
psycopg2-binary>=2.9.9
sentence-transformers>=2.2.2
groq>=0.4.2
python-dotenv>=1.0.1
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic[email]>=2.7.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
